first bag fix nest

+++ src/sessions/sessions.controller.ts
@@ -1,9 +1,20 @@
-import { Controller, Get, Param, Delete, Req } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Param,
+  Delete,
+  Req,
+  UseGuards,
+  HttpException,
+  HttpCode,
+} from '@nestjs/common';
 import { SessionsService } from './sessions.service';
 import { SessionsQueryRepository } from './sessionsQueryRepository';
 import { JwtService } from '../auth/application/jwt-service';
+import { RefreshTokenGuard } from '../auth/guards/refresh.token.guard';
+import { SessionDBType } from './dto/session.dto';
 
-@Controller('sessions')
+@Controller('/security/devices')
 export class SessionsController {
   constructor(
     private readonly sessionsService: SessionsService,
@@ -11,12 +22,23 @@ export class SessionsController {
     private jwtService: JwtService,
   ) {}
 
-  @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.sessionsQueryRepository.getSession(id);
+  @UseGuards(RefreshTokenGuard)
+  @Get()
+  async findOne(@Req() req) {
+    const result = await this.sessionsQueryRepository.getAllActiveSessions(
+      req.user.id,
+    );
+    return result.map((s: SessionDBType) => ({
+      ip: s.ip,
+      title: s.title,
+      lastActiveDate: s.lastActiveDate,
+      deviceId: s.deviceId,
+    }));
   }
 
+  @UseGuards(RefreshTokenGuard)
   @Delete()
+  @HttpCode(204)
   async deleteAllSessionsExceptOne(@Req() req) {
     const payload = await this.jwtService.getUserIdByRefreshToken(
       req.cookies.refreshToken.split(' ')[0],
@@ -27,7 +49,9 @@ export class SessionsController {
     );
   }
 
-  @Delete(':id')
+  @UseGuards(RefreshTokenGuard)
+  @Delete(':deviceId')
+  @HttpCode(204)
   async deleteSessionsByDeviceId(
     @Param('deviceId') deviceId: string,
     @Req() req,
@@ -35,6 +59,13 @@ export class SessionsController {
     const payload = await this.jwtService.getUserIdByRefreshToken(
       req.cookies.refreshToken.split(' ')[0],
     );
+    const comment = await this.sessionsQueryRepository.getSession(deviceId);
+    if (!comment) {
+      throw new HttpException({}, 404);
+    }
+    if (comment.userId !== payload.userId) {
+      throw new HttpException({}, 403);
+    }
     return await this.sessionsService.deleteSessionsByDeviceId(
       payload.userId,
       deviceId,
